back-end/requirements.txt
flask==3.0.0
peewee==3.17.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0